VC05-001 Update outputs
- Q930-002__itp is using -gnato33 so using universal integer removed
some overflow checks
- S327-020__reduced uses -gnato13 which means that there is no overflow
check inside the assertion like before, but now GNAT and GNATprove use
universal integer to evaluate 2 ** K instead of a modular type so the
assertion fails (which is an improvement). If we remove -gnato13 we get
the overflow check.
- S329-030__bound_exponentiation uses normal machine integers, so we
get here a (passing) overflow check because the frontend marks 2 ** I
with the Do_Overflow_Check flag.

Change-Id: I993d788d74ff07858c2f549b1b15fa5260018e70