Support floats in RAC
Add support for floats in RAC, to get checked counterexamples
involving directly or indirectly floats (when the code uses floats
for example, even if not directly related to the counterexample).

Fix various issues with RAC.

Add test to exercise code of RAC outside of a counterexample,
and check that it matches actual execution.