VC12-038 file-based caching
We introduce a new way to cache prover and gnatwhy3 results, via
files stored on disk. We modify the existing memcached-wrapper to
use an abstract interface, and modify the memcached client code to
implement this abstract interface. We then introduce a new
implementation for this interface based on files. Finally, the
Init_Client procedure creates the correct client based on the
commandline arguments.

* cache_client.ads
  new file to contain the abstract interface of a cache
* filecache_client.adb
  new unit to contain the file-based implementation of a cache
* memcache_client.ads
  changes to the interface to make the memcached_client derive from
  the abstract cache interface
* spark_memcached_wrapper.adb
(Init_Client): depending on the commandline, return a file-based or
  server-based client; return a classwide type.
(SPARK_Memcached_Wrapper): use dot notation for calls to cache
object; use classwide type for cache object.

Change-Id: I17ba4707f224fc67b3e1cd491de2e75b88116202