Fix crash on overlay of protected component
We support overlays of standalone objects (i.e. variables and constants)
and of subprogram parameters, but crash on overlays of (protected)
components. This patch gently rejects such overlays, which might also
cause problems when a protected function overlays a protected component
with a local variable.

[changelog]
* src/spark/spark_util.adb (Supported_Alias): Only accept stand-alone
objects, formal parameters and loop parameters.
* testsuite/gnatprove/tests/U316-010__protected/
Add test, which otherwise was causing a crash.

TN: U316-010
Change-Id: I290ff235a81279c43c0f85d4a2026b86ab8127e9