VC16-017 Fix suppression on continuation check message
When a suppressed check message has a continuation, the continuation
should be ignored instead of being added as a message in JSON.
The alternative of adding it as a suppressed message is not as good,
as this leads to a mismatch between the log in gnatprove.out and the
textual output (where only one "justified" message is output).

[changelog]
* src/flow/flow_error_messages.adb (Error_Msg_Flow): Remove unused
Last_Suppr_Str. The bug was introduced when the structure
Suppressed_Message was added instead of a simple string. Now simply
ignore continuations of suppressed messages.
* testsuite/gnatprove/tests/VC16-017__justified_check: New test.
* testsuite/gnatprove/tests/P707-026__continuation/test.out
Update output.

Change-Id: I511d6b0c6c91e44e0c3ae781083fb9d2fa30fe0b