W105-021 change the function Move_Memory in pointers library
The function Move_Memory in the library of pointers with aliasing
and separate memory now moves several memory location at once.

* include/spark-pointers-abstract_maps.ads
("="): Make equality on maps not executable so we cannot execute it and
 see that an empty map is the same as a map which is supposed to hold
 some elements.

* include/spark-pointers-abstract_sets.ads
Introduce a non executable type for sets with size 0. It can be used to
model a ghost subprogram parameter or a ghost component.

* include/spark-pointers-pointers_with_aliasing_separate_memory.ads
Replace the address array by an abstract state of addresses for
footprints. Use it as a parameter of Move_Memory.

* tests/U908-033__hidden_pointers/test.out
Changes in source location.

* tests/U908-033__hidden_pointers_separate_mem/example_recursive.adb
Add a new function merge which links together two cyclic lists using
Move_Memory.

Change-Id: Ied49f3d3a95185dbffa8f105ca507f9e622fe6f3