V930-032 Fix possible crash seen on during development
The removal of the binding for the loop iteration variable can cause
a crash when outputting values of variables involved in the check, after
a fuzzing session, as the values are taken from the context. This was
seen on test LB27-011__overflow_other_switch as part of development.

[changelog]
* src/counterexamples/ce_rac.adb: Do not remove binding when RAC reached
a check failure.

Change-Id: Id375876be33014acd592ac984266dd857e10a01b