Robustify file-based cache
For file-based caching, we write the content of the process in a temp
file, before renaming it into the cache file. It's better to create the
temp file in the same dir as the cache file, as renaming across devices
isn't always supported, and if the files are in the same dir, they
should also be on the same device.

* filecache_client.adb
(Set): first set the directory to the cache dir, then set it back