W105-004 mark function before actuals on calls
It is necessary so that a potential higher order specialization
annotation on the subprogram is found when marking the actuals.

* src/spark/spark_definition.adb
(Mark_Call): Mark function before actuals on calls.

* tests/ug__long__perm
Add an assertion. The test did not prove anymore probably due to
a change in the order in which entities are translated.