W105-004 reject specialized parameter in iterated component association
Also add a sanity checking mechanism which should ensure that the
tool cannot verify the program if we somehow miss a usage of a
specialized parameter.

* src/spark/spark_definition-annotate.adb
(Check_Higher_Order_Specialization_Annotation): Reject all references
 to specializable formals inside iterated component associations in
 contracts.

* src/why/why-gen-binders.ads
(Unit_Param): Add an optional parameter to store an Ada node in the
 binder.

* src/why/gnat2why-subprograms.adb
(Compute_Raw_Binders): Provide an Ada node for binders for specialized
 parameters so they will be stored in the Symbol_Table. This works as a
 sanity checking. If the formal is referenced, which should never
 happen, the generated Why3 code will be ill-typed.
(Compute_Guard_Formula): Ignore specialized formals. It was not
 necessary before because binders for specialized formals had no Ada
 nodes.

* tests/W105-004__iterated_comp_assoc
New test with a reference to a specialized formal in an iterated
component association.