W105-004 introduce specific module for recursive specialized calls
* src/why/why-atree-modules.ads
(M_HO_Specialization_Type): Add the module for the axiom of recursive
 functions.
(M_HO_Specializations): Add an indirection, specialization of the
 same function are grouped together in a map. This is useful when
 looking for the axiom module of recursive subprograms specializations.

* src/why/gnat2why-subprograms-pointers.adb
(Create_Theory_For_HO_Specialization_If_Needed): Adapt to the changes
 in M_HO_Specializations and create a module for the axiom of recursive
 functions.

* src/why/gnat2why-expr.adb
(Transform_Function_Call): Adapt to the changes in
 M_HO_Specializations.

* src/why/why-gen-names.adb
(Guard_Predicate_Name, Logic_Function_Name): Idem.

* src/why/gnat2why-subprograms.adb
(Generate_Axiom_For_Post): Look for the name of the axiom module in
 the specialization module if any.

* src/why/why-atree-modules.adb
(Mutually_Recursive_Modules): Exclude axioms of specializations of
 recursive subprograms.

* tests/W105-004__recursive_wrong_axiom
New test to make sure that the axiom of specialized recursive calls is
not included.