W105-004 enforce restrictions on lemmas of specializable functions
Here, we have chosen to not reject lemmas which cannot be easily
specialized with their associated function but rather to warn the
user that these lemmas will be unavailable on specializations.
The checks are delayed because we do not know in which order the
higher order specialization and automatic instantiation annotations
will be analyzed.

* src/spark/spark_definition-annotate.adb
(Delayed_Checks_For_Lemmas): Set of lemma procedures with automatic
 instantiation that need to be checked for consistency with respect
 to higher order specialization annotations.
(Higher_Order_Spec_Annotations): Add a mapping from functions
 annotated with Higher_Order_Specialization to a potential set of
 lemmas that should be automatically instantiated when the function is
 specialized.
(Check_Automatic_Inst_And_HO_Specialization_Compatibility): Check that
 lemmas associated to a function with higher order specialization can
 be specialized with the function. If it is not the case, emit a
 warning. Store compatible lemmas in the Higher_Order_Spec_Annotation
 map.
(Check_Automatic_Instantiation_Annotation): Store lemmas in the
 Delayed_Checks_For_Lemmas map to be checked later. Indeed, we do not
 know in which order the higher order specialization and automatic
 instantiation annotations will be analyzed.
(Do_Delayed_Checks_On_Pragma_Annotate): Go over the set of lemmas with
 automatic instantiation. For those associated to functions with higher
 order specializations, check that the lemma can be specialized and
 fill the mapping in Higher_Order_Spec_Annotations.

* tests/W105-004__warnings_on_lemmas
New test with warnings on lemmas which are not easily specializable.