Use subprogram type when borrow checking simple return statements
When borrow checking expression of a simple return statement, it is
better to look at the type of the enclosing function, because it might
differ from the type of the returned expression, e.g. function might
be of an anonymous access type, while expression might be of a named
access type.

[changelog]
* src/why/gnat2why-borrow_checker.adb
(Check_Simple_Return_Expression): Use type of the subprogram, not of the
expression.

TN: VB02-016