W125-017 Return partial view of entities to proof
This patch fixes a problem where borrow checker expects an entity
and is returned a Magic_String through Get_Variables_For_Proof.
The case was a deferred constant whose full view is not in SPARK,
and flow returned the magic string without checking if the partial
view was in SPARK.

* src/flow/flow_generated_globals-phase_2.adb
(Expand_Abstract_State): Return the partial view of an entity if said
entity is not in SPARK but its partial view is.

* src/flow/flow_utility.adb
(Path_To_Flow_Id): Use Unique_Entity to get only one view of entities.

* testsuite/gnatprove/tests/W125-017__flow_partial_view_deferred_constant/
Add test.