Cleanup computing of tag checks for protected entries
The idea of tag checks applies only to dispatching subprogram calls
and not to "dispatching entry calls" (because these are not possible
in Ada). This is now reflected in the code itself.

Cleanup related to switching to unexpanded expression functions;
behavior is unaffected, but the performance should be slightly improved.

[changelog]
* src/why/gnat2why-expr.adb (Compute_Tag_Check): Add contract; adapt
caller; remove now unnecessary handling of protected entries.

no-tn-check