V927-049 remove useless axiom for constants
The axiom module for constants was included in the VC module for the
enclosing subprogram, even though this is not needed (the value is
"assigned" to the constant during the VC generation). We don't add these
modules when adding the axiom modules.

* why-inter.adb
(Add_Axiom_Imports): don't add axioms for local constants
* gnat2why-expr.adb
(Assignment_Of_Obj_Decl): no need to check for RTE in static constants
* W105-004__count_on_sets: accept regression