W130-013 update parentheses in pretty-printed counterexamples
Expressions that are parenthesized in the source code are now also
parenthesized in the pretty-printed counterexamples; those which are not
parenthesized in the source code are not parenthesized in pretty-printed
counterexamples.

We could argue which approach is better, but at least from the
implementation point of view the new display is simpler and consistent.

Also, it would be troublesome to revert it without causing diffs in
slocs, because both pretty-printing and locating messages uses the same
routine First_Sloc.