W130-013 move code for fixing parentheses from backend to frontend
Expressions with parentheses are notoriously problematic to
pretty-print. In GNATprove we had a post-processing fix for them, but it
is better to have this fix in the GNAT frontend repository and apply it
for CodePeer as well.

* src/spark/spark_util.adb
(Count_Parentheses, Fix_Parentheses): Move to frontend.
(String_Of_Node): Don't fix parentheses.