W210-019 precise in_range predicate on derived access-to-subprogram
Minor, reuse the in_range predicate of the parent on derived
access-to-function types and subtypes. This predicate is used to
provide the postcondition of the function in an axiom in a sound
way. This is fine as contracts cannot be redefined on derived types
and subtypes.

* src/why/gnat2why-subprograms-pointers.adb
(Declare_Access_To_Subprogram_Type): Use the in_range predicate of the
 parent for derived types and subtypes.