V927-049 Detection of cycles in proof axiom modules
We maintain a map of modules to entities, where an entry m -> e
signifies that the (axiom) module m depends on the proof of the entity
e. We then use this map together with the module inclusions for VC
modules to compute a dependency graph between entities to detect cycles
(for example, the VC module for A uses the axiom of B, and the VC module
of B uses the axiom of A). We issue a hard error in this case.

* flow_generated_globals-phase_2.adb
(Add_Edges): use Enclosing_Unit_SPARK
* spark_util.adb
(Enclosing_Unit): take into account wrapper packages
* gnat2why-decls.adb
(Translate_Constant_Value): register that the axiom module for the
constant depends on the proof of the enclosing subprogram
* gnat2why-driver.adb
(Translate_CUnit): check for cycles after having translated all units
* gnat2why-subprograms-pointers.adb
(Declare_Theory_For_Access_If_Needed): register that the axiom module
for 'Access depends on proof of the corresponding subprogram
* gnat2why-subprograms.adb
(Emit_Post_Axiom): register that the axiom module for a subprogram
depends on its own proof
(Generate_Subprogram_Completion): register that the dispatch axiom
module for a subprogram depends on the proof of the subprogram
* why-inter.adb
(Close_Axiom): build the entity graph to detect cycles based on the
  included axiom modules
(Check_Safe_Guard_Cycles): new check for cycles in the entity graph

Testsuite:
P518-021__frame_for_max: accept a regression, likely due to removal of
  constant axiom
Other: new messages