W130-013 accept extra space in pretty-printed counterexamples
These extra spaces occur because the machinery for detecting parentheses
first descends the AST using Original_Node and then climbs back using
Parent, so it can't reach the initial node.

Years ago we experienced the same problem in flow analysis for aliasing
checks with inlining-for-proof and fixed it there. Here the fix is not
that easy to implement, while the occasionally garbled output is
acceptable.