W222-010 fix regression of mod for Storage_Offset
Calls to "mod" in System.Storage_Elements, which takes an address and a
Storage_Offset, now appear as "mod" operators in the tree instead of as
function calls. We move the special casing of this operation in a new
subprogram called Transform_Hardcoded_Operation.

We don't easily have access to the entity of the mod operation, so we
remove the continuation message for the precondition, which was attached
to the entity.

spark_util-hardcoded.adb
(Is_System_Address, Has_Stoele_Offset): new functions to check if the
mod operation is the one we are looking for
spark_util-subprograms.adb
(Is_Unchecked_Deallocation_Instance): this function was crashing when
calling it on an inherited primitive operator
gnat2why-expr.adb
(Transform_Expr): before translating a rem/mod operation, check if it is
  hardcoded
why-gen-hardcoded.adb
(Transform_Hardcoded_Function_Call): remove code for precondition of mod
operation from here, as it is no longer treated as a function call
(Transform_Hardcoded_Operation): new subprogram to hold the code for the
hardcoded mod operation.