Remove redundant assumption about globals and dispatching calls
In general, we want the overriding operation to be "not worse" than the
overridden one. We don't specifically enforce this for any of the
tasking-related properties, but the existing rules about class-wide
Global and Depends contracts are enough to at least enforce this for
accesses to (unsynchronised) globals.

* docs/ug/en/source/how_to_use_gnatprove_in_a_team.rst
[SPARK_OVERRIDING_AND_TASKING]: Remove redundant rule.

TN: VA14-049