W309-008 fix gnatprove redirect of stderr
The Call_With_Status function had a parameter to redirect the output,
and System.Os_Lib only supports that for stdout, it also enabled
redirection of stderr to stdout. But this parameter was never used
(always used with the default argument that redirects to stdout).
Removing this parameter, and removing the redirection.

* call.adb
(Call_With_Status): remove unused Output file descriptor argument, and
never redirect stderr
* ipstack: add redirection to Makefile sort