W315-018 pass the set of fields with missing values for aggregates
When translating values of an Ada aggregate into Why3, some values
might be missing on variant records (those for other variants). These
components are initialized by default as their value does not matter.
Compute the set of missing components when translating the aggregate
value and pass it to the code computing the Why3 aggregate instead of
trying to guess which fields are going to be missing.

* src/why/gnat2why-expr.adb
(Transform_Record_Component_Associations): Create a set of missing
 components from the associations and return it.
(Transform_Delta_Aggregate): Ignore the missing components set as delta
 aggregates are always partial updates.
(Transform_Expr): Compute the set of missing components for aggregates
 and pass it along.

* src/why/why-gen-records.adb
(Generate_Associations_From_Ancestor): Update the set of missing
 components to remove those coming from the ancestor.
(New_Ada_Record_Aggregate): Take the set of missing components as
 parameter.

* tests/W315-018__aggregates
File test.