UC02-037 refactor handling of block statements
Because loop invariants are allowed to occur in nested block
statements, GNATprove sometimes needs to flatten a list of statements
so that declarations and statements of nested blocks are all considered
together as a big sequence. This is not very compatible with handling
of local borrowers and memory leaks which are checked at the end of the
block, and not compatible at all with exception handlers.
Only flatten block statements when necessary, ie. to handle loop
invariants which are nested in these blocks. Remove the workaround used
to handle local borrowers and memory leaks which used to include the
block at the end of the inlined statements as a placeholder.

* src/spark/spark_definition.adb
(Check_Loop_Invariant_Placement): No longer flatten the statements.
 Handle nested blocks instead.

* src/why/gnat2why-subprograms.adb
(Get_Pre_Post_Pragmas): Use directly the declaration list, no need to
 flatten it.

* src/spark/spark_util.adb
(Get_Flat_Statement_And_Declaration_List): Moved to
 gnat2why-expr-loops.

* src/why/gnat2why-expr-loops.adb
(Get_Flat_Statement_And_Declaration_List): Only inline block statements
 if they contain a loop invariant or variant. Local borrowers and
 objects which need reclamation should never occur in an inlined block.
(Transform_Loop_Body_Statements): Remove the special handling of blocks
 as placeholders.

* src/why/gnat2why-expr-loops-inv.adb
(Process_Loop_Statement): Removed, inlined at use. Because of this
 inlining, the parameter Inv_Objects is no longer needed for
 Process_Statement and Process_Statements_List.
(Get_Loop_Writes): Inline Process_Loop_Statement. The statements are
 flattened to more easily collect objects declared just before the
 invariant.
(Process_Statement): Handle block statements which are not inlined.
 Inline Process_Loop_Statement.
(Process_Statements_List): Do not flatten the list of statements.