Fix handling of parameters of E_Subprogram_Type entities in proof
Remove a leftover code from early task types support, which was meant to
filter the implicit task parameter when translating the task type
itself. We no longer translate task types with this routine and this
code was wrongly ignoring binders for parameters of access-to-subprogram
types.

[changelog]
* src/why/gnat2why-subprograms.adb (Generate_Subprogram_Program_Fun):
Insert parameters of an E_Subprogram_Type entity into map.

TN: W322-033