Use vertex context that is carried while constructing CFG
If we ever add a component to the Vertex_Context record, we expect its
value to be used when creating vertex attributes. The currently used
"No_Vertex_Context with delta ..." need to be modified after adding a
new component. That was never intended and is difficult to maintain.

Instead, it seems better to "pass the context by default" and only use
"with delta" where necessary.

[changelog]
* src/flow/flow-control_flow_graph.adb
(Do_Simple_Return_Statement): Use existing context; the precondition
makes sure that its In_Nested_Package component is false.
(Do_Subprogram_Or_Block): When processing entry bodies the context is
null, so we can just use it.
(Create, Reclaim_Borrower): Fix whitespace.

no-tn-check