Improve "possible fix" message regarding bounds and types
Improve the "possible fix" message in the case of a failing range or
overflow check:
- When only the lower/upper bound is not proved, express the fix in
terms of that bound.
- When the expression and the bounds are not in the same type, convert
when possible one of the terms of the comparison to the other type,
where the larger type is chosen for the comparison.

Add test 160__possible_fix for that improvement.

Update test outputs with more precise possible fix messages.