Add test with inconsistent flow for raise statement and expression
These tests have semantically similar raise_statement and
raise_expression accompanied by "with S", where "S" is uninitialized.

For raise_statement the path leading to it is pruned, so we don't
complain on S being uninitialized. For raise_expression we don't prune
and so we complain. Same for variables leading to the raise_expression.

* testsuite/gnatprove/tests/185__flow_raise/
Add; for func.adb we want check on "S" to become warning that it is not
initialized, like for proc.adb.