Improve proof of termination of loops over iterable constants
When flow analysis detects a loop over an expression path whose
root object is a variable that is not modified during the loop,
it automatically proves its termination.

[changelog]
* src/flow/flow-control_flow_graph.adb
(Do_Loop_Statement): Move the creation of Loop_Writes after processing
the loop statements. Termination_Proved needs this information to detect
when the root object of an expression path is not modified during the loop.
In this case, termination of the loop can be safely assumed.