No longer raise warnings on objects annotated with external names
This more or less reverts the changes done for eng/spark/spark2014#171.
No warnings are raised for objects annotated with External_Name or
Link_Name. The assumptions are updated to reflect this change, but
the requirements stay the same.

[change_log]
* en/source/how_to_use_gnatprove_in_a_team.rst
Update the related assumptions.

* src/common/vc_kinds.adb
Update the message desciption of warnings.

* src/spark/spark_definition.adb
* src/why/gnat2why-expr.adb
Do not emit warnings on objects annotated with External_Name or
Link_Name.

* src/spark/spark_util.ads
(Has_Address_Or_Name): Removed.

* tests/*/test.out
Warnings are no longer emitted.