Detect implicit terminating annotations in phase 1
Functions and automatically instantiated lemmas were not detected
to have an implicit terminating annotation in phase 1. This lead
to spurious checks when such subprogram was called from a procedure
not annotated with a terminating annotation.

* src/flow/flow_generated_globals-partial.adb
(Do_Preanalysis): Detect implicit terminating annotations along with
explicit terminating annotations to store the information in the ALI
files.

* testsuite/gnatprove/tests/216__flow_term_func_callee/
* testsuite/gnatprove/tests/P715-019__flow_subprogram_termination/
Remove spurious checks.