Fix warning code for some warnings about calls
Some of the warnings in Mark_Calls were only issued if
Emit_Warning_Info_Messages is true, but some others were issued
unconditionally. This seems to be a mistake, and we now check
Emit_Warning_Info_Messages for all those warnings.

* spark_definition.adb
(Mark_Call): rearrange code so that all warnings are protected by
  Emit_Warning_Info_Messages