Add tests with raising of exception being recognized as an effect
Raising an unhandled exception of one kind or another, or not raising it
at all, should be recognized as an effect, even though it is not
represented by the Depends contract. Currently this causes a spurious
warning in the callee (see test Effect1), even though at the caller
things work as expected (see test Effect2).

Issue: eng/spark/spark2014#130