Add tests with reclaiming of borrows on the paths to handlers
When raising exception that will be handled in the current block we
should not reclaim borrowers declared in this very block. However, when
handling exception that will be handled in an enclosing block, we should
reclaim borrowers that go out of scope.

Those two cases are exercised in the added tests, both for raise
statements and for calls that raise exceptions (currently those two
variants duplicate code in flow analysis). The case 3 merely combines
cases 1 and 2.

Unfortunately, it is a bit hard to write a meaningful test for this
issue, so the added tests are rather intended for manual inspections of
CFGs.

[changelog]
* testsuite/gnatprove/tests/130__flow_exception/
Add tests and update output.

Issue: eng/spark/spark2014#130