New meta remove_unused:remove_constant
Most drivers use the remove_unused_keep_constants transformation. A bug
in the implementation of this transformation resulted in some constants
being removed despite the "keep_constants" promised by the name.
It turns out, this was beneficial for the ada__model library since it
removed the constants named "dummy" in the library. A recent fix to this
bug in Why3 caused many counterexamples to be lost.
To overcome this, a new meta directive is introduced to mark constants
that must be deleted even by remove_unused_keep_constants. This meta is
now added to all dummy constants.