Fix untangling of class-wide conversions in flow
When untangling a record type conversion, we were collect valid fields
of the target type (we store them in a local variable "Valid_To_Fields")
and use them to pick input/output pairs that can be precisely
represented (we store them in a local variable "M"). For other pairs, in
particular an extra fields in a "downcast" class-wide conversion, we
were wrongly adding a mapping from the "extension field" to the valid
fields; now we add a mapping to the inputs of the missing fields.

* src/flow/flow_utility.adb
(Untangle_Record_Assignment): Collect extra inputs into the extension
part.
* testsuite/gnatprove/tests/248__flow_classwide/
Add minimal reproducer; it no longer causes an assertion failure.

Issue: eng/spark/spark2014#248