Adapt test after a frontend fixes for private views in instances
One test was supplying an Integer "42" to a generic instance that
expected a System.Address. Now this is rejected by the frontend and
requires an explicit conversion.

This patch adapts this test to be legal in both Ada and SPARK.  (The
System'To_Address could be used as well, but causes an unrelated
frontend crash.)

* testsuite/gnatprove/tests/W125-017__flow_partial_view_...
Adapt.

TN: W125-017
TN: W112-018