Fix untangling of overlaying record objects
Reference to a selected component of an overlaying record object, like:

   X : Record_Type with Address => Other_Object'Address, Import;
   ... := X.Component;

is handled like a reference to all components of the underlying object.

This was already implemented in Get_Assignment_Target_Properties and
few other routines, but not in Untangle_Record_Fields.

[changelog]
* src/flow/flow_utility.adb
(Untangle_Record_Fields): Handle record overlays.

Issue: eng/spark/spark2014#260