Crash on unknown variables but not on wrongly generated globals
Partly revert recent change that introduced a hard crash when unknown
variable appears in a subprogram with generated globals. Now we only
have a hard crash when the variable is local to the subprogram.

[changelog]
* src/flow/flow-analysis-sanity.adb
(Check_All_Variables_Known): Raise exception on unknown local variables,
but fail with assertion on unknown global variables.

Issue: eng/spark/spark2014#207