Move the tag update to the defining axiom of array aggregates
If an array aggregate contains tagged components, then the tag of the
values provided for the components is set to the expected component tag
when the value is assigned into the array. Do the update reflecting
this assignment inside the defining axiom of the aggregate function
instead of on the function parameter.