Fix dependencies for particular exception being raised by a call
In CDG, or control dependency graph, we added edges that represent
raising an exception by a call being dependent on the call inputs.

With this patch we add more edges like that, to represent that choice of
a particular exception being raised also depends on the call inputs.

[changelog]
* src/flow/flow-control_dependence_graph.adb
(Create): Move and extend comment; add edges also to the havoc vertex;
skip the call exception vertex when navigating from havoc vertex to
input vertices.
* src/flow/flow-interprocedural.adb
(Find_Parameter_Vertex): Skip the call exception vertex when navigating
from havoc vertex to input vertices.
* testsuite/gnatprove/tests/130__flow_exception/test.out
Spurious checks are gone.