Remove redundant message on read global output
For global output variables that are read before being
initialized, we no longer emit a message saying that the variable
is not initialized because we already emit a message stating that
the variable is not an input in the global contracts.

[changelog]
* src/flow/flow-analysis.adb
(Emit_Check_Message): For global output variables, we no longer
emit a message saying that the variable is not initialized.

 * testsuite/gnatprove/tests/*
 Updated tests.

 Issue: eng/spark/spark2014#77