Use markers for all checks in SPARK testsuite
This completes the use of markers in the SPARK testsuite, so that all checks
have corresponding markers with the expected value FAIL, indicating that
GNATprove fails to prove a false property (or one that should not be provable).
This is what matters for soundness of the tool, so that's what we are checking.

This concerns the 57 kinds of checks as identified in vc_kinds.ads in the
sources of GNATprove, leaving out the additional 5 special checks for warnings,
which are not meant to be sound. There are 56 markers in total, as checks
VC_UC_SOURCE and VC_UC_TARGET are tested together (whether the source and target
types of an Unchecked_Conversion or overlay are suitable).

In particular, we use separate markers instead of a shared one as before, for:
- VC_OVERFLOW_CHECK and VC_FP_OVERFLOW_CHECK
- VC_PREDICATE_CHECK and VC_PREDICATE_CHECK_ON_DEFAULT_VALUE
- VC_INVARIANT_CHECK and VC_INVARIANT_CHECK_ON_DEFAULT_VALUE
- VC_RESOURCE_LEAK and VC_RESOURCE_LEAK_AT_END_OF_SCOPE

Only one test was added, 14_iso26262_qualification__weaker_classwide_pre, for
VC_WEAKER_CLASSWIDE_PRE, using code from adacore_u which is put here at the right
level for being picked for the source tab verification (which is only done for
sources at top level).

Issue: eng/cert/projects/iso26262-nvidia-runtime-qualification#14