Special case concat_singleton_left when a_first = b_first
Add a specialization of the general axiom for concatenation of an
element to the left of a 1-dimensional array for the case where
a_first and b_first are the same. This is a common enough case for
arrays with a fixed lower bound.

[changelog]
* share/spark/theories/_gnatprove_standard.mlw
(Array_1__Concat): Replace the axiom for concat_singleton_left by
 3 axioms: one giving the value of the first element, and 2 giving
 the value of the other elements, with and without the additional
 constraint that a_first = b_first.

* tests/276__concat
New test.

* tests/O702-009__coq_realizations/realize/Array__1__Concat.v
Update realization.