Fix detection of boundary function calls in type invariants
Not every boundary function call is disallowed in a loop invariant for a
given type T. It is only a problem if the function has an input with a
part of type T. Refine the test.

[changelog]
* src/flow/flow-analysis-sanity.adb
(Check_Type_Aspects): Only emit the check on boundary function calls if
 the callee has formal parameters with a part of type T.

* src/spark/spark_util-types.adb
(Traverse_Subcomponents): Generic function which applies test to all
 subcomponents of a type until one is found on which a function Test
 returns Pass.
(Traverse_Access_Parts): Reuse Traverse_Subcomponents.
(Has_Subcomponents_Of_Type): Return True if a type has subcomponents of
 a given type or a derived type thereof.

* lib/python/test_support.py
Add a marker for checks for calls to boundary functions in type
invariants.

* tests/Q613-018__flow_type_invariants/test.out
* tests/R417-042__flow_type_aspects/test.out
* tests/R514-013__invariant/test.out
* tests/R514-018__type_inv/test.out
* tests/TU__type_invariant__legal/test.out
Remove spurious checks.

* tests/128__type_invariants
New test.