Examine hidden globals of boundary function calls
Reject boundary function calls if their globals are not known by
Entity_Id. The actual illegality will most likely be illegal anyway due
to circular dependencies, so this patch is rather quite conservative,
but it is the easiest way to stay on the safe side.

[changelog]
* src/flow/flow-analysis-sanity.adb
(Check_Type_Aspects): Reject boundary calls if we can't examine types of
their globals.