Do not look for Is_Moved predicates on limited types
Since limited types cannot be moved, the Move function and Is_Moved
predicate are not generated for them. When checking absence of momory
leaks, the Is_Moved condition has to be computed on the fly.

[changelog]
* src/why/gnat2why-expr.adb
(Compute_Is_Moved_Property): Do not use the predicate of limited types.

* tests/294__limited
File test. Counterexamples are disabled as they cause cvc5 processes to
runaway.