Quantification and iteration with Iterable cannot dispatch
Quantification and iteration with Iterable cannot be used on a
classwide type (it is rejected in the frontend). As a result, no
dispatching can occur inside such iterations.

* src/spark/spark_definition.adb
(Mark): Add an assertion that the name of the iterator specification
 does not have a classwide type.

* src/spark/spark_util-types.adb
(Declares_Iterable_Aspect): Return False on classwide types.

* src/why/gnat2why-expr.adb
(Check_Type_With_Iterable): Remove special case for Iterable with
 dispatching primitives and add a forgotten conversion.

* tests/300__dispatch_iterable
File test. An assertion is commented as it crashes in flow analysis for
now.