Move code of Is_Statically_Disabled to GNAT frontend
As the frontend also needs to detect cases of "statically disabled
conditions which evaluates to a value" as described in SPARK UG 7.3.2,
we move this code to GNAT frontend in Exp_Util.

Prefix all uses by Exp_Util rather that using the unit, as we rarely
reference expansion functions, hence it is better to make this visible,
as expansion may in some cases require a different form of AST (which
is not the case here).

Issue: eng/toolchain/gnat#407
Depends-On: https://gitlab.adacore-it.com/eng/toolchain/gnat/-/merge_requests/215