Improve --checks-as-errors switch
This change allows users to disable --checks-as-errors after having
switched it on. Now, an argument "on" or "off" needs to be provided.
This is an incompatible change.

* configuration.adb
- change Checks_As_Errors type to String, introduce new synthetized
  variable for Checks_As_Errors that holds the boolean
(Parse_Switches): specify argument for --checks-as-errors switch
(Postprocess): set boolean Checks_As_Errors based on command-line string
  argument
* gnatprove.adb
(Generate_SPARK_Report): use synthetized variable instead of directly
  the command-line one
* help.txt: adapt doc for command-line switches
* how_to_run_gnatprove.rst: adapt documentation
* QB03-022__exit_status/test.py: adapt usage of command-line switch