Access types can designate parts with Relaxed_Initialization
Calling Contains_Relaxed_Init_Parts in marking might lead to a
Program_Error if the input type has access subcomponents designating
unmarked private or incomplete types. This was deemed acceptable for
now as it should not happen often. If we find a need, it will be
necessary to defer the relaxed initialization checks on such types to
the end of the marking.