Support for-of loops and quantified expressions in RAC
Lack of support for for-of loops and quantified expressions lead to
the loss of counterexamples. Now improved.

Loops over a rage and over an array (the for-of loops we intend to support
now) are treated similarly in the renamed Iterate_Scheme_Spec.

In addition to the dedicated test, a dozen tests have improved outputs
as a result.