New limitation for null array subaggregate
Null array subaggregates of a (necessarily multidimensional)
array aggregate are now rejected if the aggregate contains
several associations (whether named or positional).

This case is useless in practice (can always factor the cases for
each dimension if the aggregate is the null one from
some dimension on, making a simpler aggregate),
and difficult to handle currently for bound checking
due to the front-end not providing all the required bounds.